Supports trigger.io's storage APIs
Setting "useTriggerIO" to true in the settings will make TaffyDB use
trigger.io's storage APIs. This means that on mobile devices, it will
use the native storage capabilities, thus getting around the storage
size limits of localStorage.

Due to the asynchronous nature of Trigger.io's APIs, it is best to
initialise your Taffy() instance, set the "useTriggerIO" setting, then
insert data to ensure it starts writing to the storage immediately.

When reloading a pre-existing database, there will be a delay before it
is completed. For now, I am using a setTimeout(doSomethingWithDB,2000)
call after the initial call to .store(). A better solution will
hopefully be created shortly.